refactor: separa pagination para componente
