+++ src/app/fichas/RecordsTable.tsx
@@ -1,6 +1,7 @@
 import { MoreVertical } from "lucide-react";
 import { useState } from "react";
 import { DeleteConfirmModal } from "@/components/DeleteConfirmModal";
+import { Pagination } from "@/components/Pagination";
 import { deletePaciente } from "@/services/paciente.service";
 
 type Record = {
@@ -159,50 +160,6 @@ const RecordRow = ({ record, onRefresh }: { record: Record; onRefresh: () => voi
         <ActionsDropdown recordName={record.nome} onRefresh={onRefresh} />
       </td>
     </tr>
-  );
-};
-
-const Pagination = ({
-  total,
-  currentPage,
-  onPageChange,
-}: {
-  total: number;
-  currentPage: number;
-  onPageChange: (page: number) => void;
-}) => {
-  const pageSize = 10;
-  const totalPages = Math.ceil(total / pageSize);
-
-  if (totalPages <= 1) return null;
-
-  const pages = [];
-  for (let i = 1; i <= totalPages; i++) {
-    pages.push(i);
-  }
-
-  return (
-    <div className="flex items-center justify-between mt-4">
-      <p className="text-sm text-[var(--text)]">
-        {total} {total === 1 ? "registro encontrado" : "registros encontrados"}
-      </p>
-      <div className="flex gap-2">
-        {pages.map((page) => (
-          <button
-            key={page}
-            type="button"
-            onClick={() => onPageChange(page)}
-            className={`px-3 py-1 rounded-full font-semibold ${
-              page === currentPage
-                ? "bg-[var(--primary)] text-white"
-                : "bg-[var(--white)] text-[var(--primary)] hover:bg-gray-100"
-            }`}
-          >
-            {page}
-          </button>
-        ))}
-      </div>
-    </div>
   );
 };
